requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
huggingface-hub
python-dotenv